[fix] Moved UploadButton from containers to components
